apps/vanna/requirements.txt
fastapi
uvicorn[standard]
httpx
asyncpg
pydantic
python-dotenv
